feat: zod refactor
