+++ package.json
@@ -13,7 +13,7 @@
   "dependencies": {
     "@hookform/resolvers": "^3.9.0",
     "@reduxjs/toolkit": "^2.2.7",
-    "@technosamurai/techno-ui-kit": "^0.0.9",
+    "@technosamurai/techno-ui-kit": "^0.0.10",
     "clsx": "^2.1.1",
     "next": "14.2.5",
     "next-redux-wrapper": "^8.1.0",
